Data of safety meta-analysis.

Safety meta-analysis of aggregated datasets rely on 2 major compounds: study-level data, and event-level data. During the data management process, event-level data is grouped according to safety terms, and becomes term-level data. The study-level data begin with s_.

s_data study-level data, including total number of subjects in each study arm, and other aggregated features (median age, median follow-up)

The event-level data begin with e_.

e_data_src* event-level dataset from a specific source (src1, src2)

e_data event-level dataset, where all sources have been reconciled.

The term-level data begin with t_.

t_groups A named list. Names are the terms, contents are the events composing the term.

tnsc_data term-level dataset, also joined with some study characteristics (subgroups).

Usage

data(e_data)

e_data_src1

e_data_src2

t_groups

s_data

tnsc_data

Format

An object of class data.frame.

An object of class data.frame with 4 rows and 5 columns.

An object of class data.frame with 3 rows and 5 columns.

An object of class list of length 2.

An object of class data.frame with 3 rows and 4 columns.

An object of class data.frame with 4 rows and 8 columns.

Source

None

References

There is none

Examples

data(e_data)
e_data[e_data$study_id == "NCT1", ]
#>   source study_id              event_name n_event_arm1 n_arm1
#> 1   src1     NCT1   myocardial infarction            1     20
#> 2   src1     NCT1 acute coronary syndrome            2     20