Group events of a common term group

Allows for prioritization of event data, aggregating multiples terms under a group term (see integrative and conservative methods).

Usage

group_events(
  .data,
  t_groups,
  n_event,
  n_arm = NULL,
  event_name,
  study_id,
  src = "src",
  method = c("conservative", "integrative"),
  add_zero = FALSE
)

Arguments

.data

An event data.frame (see details)

t_groups

Named list of terms. Names are terms, contents are event names (see description).

n_event

Character vector, column with number of events in study arms (e.g. placebo...).

n_arm

Optional, character vector, column with number of patients in study arms.

event_name

Character string, column of events name.

study_id

Character string, column of main study ID

src

Character vector, column with source sorting sort_sources.

method

"integrative" or "conservative" (see details).

add_zero

Do you wish to count O events for study(ies) without any events of a term? (see description)

Value

Special attributes of the output data.table are

group_method (conservative or integrative)

group_term

event

n

They are used in subsequent functions to create a single dataset when there are multiple arms.

Details

See vignette("basic_workflow", package = "safetymeta") for details on "event" and "term" definition. With this function, you can group multiple events corresponding to a term. A t_groups list provides one or more terms as its name(s). Integrative method means counts for each event of a term will be summed up together, while conservative means only the event with the highest count will be selected. add_zero is used when an event is not reported for a study (e.g., the event line is NOT in the dataset). The principle is that since all serious events must be reported, the absence of reporting is equal to reporting 0 events. This is not true when safety data is not available at all (i.e., when there is no data for such event in any available sources). Thus, these specific studies must be excluded prior to using group_events. Some total counts are not homogeneous for a given study, depending on the reaction (notably found for all cause mortality reaction). The maximum count will be retained for no line study(ies), (when add_zero performs). group_events assumes events in .data[["events"]] are LOWER cased.

Examples

# You are very much advised to check string case over t_groups
# and .data[["event_name"]] col. All should be lower cased +++

source_order <- c("src1", "src2")

# set src before using group_events
e_data <-
  e_data |>
  sort_sources(
    method = c("source_list"),
    source_name = "source",
    source_list_order = source_order
  )

# The integrative method
 e_data |>
  group_events(
    t_groups = t_groups,
    n_event = "n_event_arm1",
    n_arm     = "n_arm1",
    event_name = "event_name",
    study_id   = "study_id",
    method = "integrative"
  )
#> # A tibble: 4 × 6
#>     src study_id term_name             n_event_arm1 n_arm1 details_n_event_arm1 
#>   <int> <chr>    <chr>                        <dbl>  <dbl> <chr>                
#> 1     1 NCT1     myocardial_infarction            3     20 myocardial infarctio…
#> 2     1 NCT2     myocardial_infarction            5     27 myocardial infarctio…
#> 3     1 NCT2     skin                            10     27 eczema (10)          
#> 4     2 NCT3     skin                            20     50 pruritus (20)        

# Contrasting with the conservative method

 e_data |>
  group_events(
    t_groups = t_groups,
    n_event = "n_event_arm1",
    n_arm     = "n_arm1",
    event_name = "event_name",
    study_id   = "study_id",
    method = "conservative"
  )
#> # A tibble: 4 × 6
#>     src study_id term_name             n_event_arm1 n_arm1 details_n_event_arm1 
#>   <int> <chr>    <chr>                        <dbl>  <dbl> <chr>                
#> 1     1 NCT1     myocardial_infarction            2     20 myocardial infarctio…
#> 2     1 NCT2     myocardial_infarction            5     27 myocardial infarctio…
#> 3     1 NCT2     skin                            10     27 eczema (10)          
#> 4     2 NCT3     skin                            20     50 pruritus (20)