Wrapper around meta::metabin

Compute a meta analysis of binary events, and allows for multiple subgroup analyses. event_colnames is similar to event.e and event.c in meta::metabin, and n_colnames is similar to n.e and n.c, but they use character strings as inputs. The *_colnames args also cover more usecase to prevent metabin failure, in particular when there are missing data. by_var is different from byvar in metabin (by_var is a character string). The same is true for stud_lab.

Usage

metabin_wrap(
  data,
  event_colnames,
  n_colnames = paste0("n_", event_colnames),
  by_var,
  stud_lab,
  ...
)

Arguments

data

Data.frame containing study information

event_colnames

List of paired experimental - control column names with event counts

n_colnames

n counts, see Details.

by_var

Character vector, subgroup variables

stud_lab

Character string, id col for each study

...

Additional arguments to be passed to metabin (method, sm...).

Value

meta::metabin models, with very slight differences (subset and call parameters).

Examples

tnsc_data |>
  dplyr::filter(term_name == "myocardial_infarction") |>
  metabin_wrap(
    event_colnames = c("event_e", "event_c"),
    n_colnames = c("n_e", "n_c"),
    by_var = c("byvar1", "byvar2"),
    stud_lab = "study_id"
  )
#> $byvar1
#> Number of studies: k = 2
#> Number of observations: o = 96 (o.e = 47, o.c = 49)
#> Number of events: e = 11
#> 
#>                          RR           95%-CI    z p-value
#> Common effect model  2.7776 [0.7850; 9.8284] 1.58  0.1131
#> Random effects model 2.7683 [0.7819; 9.8007] 1.58  0.1144
#> 
#> Quantifying heterogeneity:
#>  tau^2 = 0; tau = 0; I^2 = 0.0%; H = 1.00
#> 
#> Test of heterogeneity:
#>     Q d.f. p-value
#>  0.02    1  0.8865
#> 
#> Results for subgroups (common effect model):
#>              k     RR            95%-CI    Q I^2
#> byvar1 = a   1 3.1500 [0.3565; 27.8310] 0.00  --
#> byvar1 = b   1 2.5926 [0.5490; 12.2427] 0.00  --
#> 
#> Test for subgroup differences (common effect model):
#>                   Q d.f. p-value
#> Between groups 0.02    1  0.8865
#> 
#> Results for subgroups (random effects model):
#>              k     RR            95%-CI tau^2 tau
#> byvar1 = a   1 3.1500 [0.3565; 27.8310]    --  --
#> byvar1 = b   1 2.5926 [0.5490; 12.2427]    --  --
#> 
#> Test for subgroup differences (random effects model):
#>                   Q d.f. p-value
#> Between groups 0.02    1  0.8865
#> 
#> Details on meta-analytical method:
#> - Mantel-Haenszel method (common effect model)
#> - Inverse variance method (random effects model)
#> - Restricted maximum-likelihood estimator for tau^2
#> 
#> $byvar2
#> Number of studies: k = 2
#> Number of observations: o = 96 (o.e = 47, o.c = 49)
#> Number of events: e = 11
#> 
#>                          RR           95%-CI    z p-value
#> Common effect model  2.7776 [0.7850; 9.8284] 1.58  0.1131
#> Random effects model 2.7683 [0.7819; 9.8007] 1.58  0.1144
#> 
#> Quantifying heterogeneity:
#>  tau^2 = 0; tau = 0; I^2 = 0.0%; H = 1.00
#> 
#> Test of heterogeneity:
#>     Q d.f. p-value
#>  0.02    1  0.8865
#> 
#> Results for subgroups (common effect model):
#>              k     RR           95%-CI    Q  I^2
#> byvar2 = a   2 2.7776 [0.7850; 9.8284] 0.02 0.0%
#> 
#> Test for subgroup differences (common effect model):
#>                   Q d.f. p-value
#> Between groups 0.00    0      --
#> 
#> Results for subgroups (random effects model):
#>              k     RR           95%-CI tau^2 tau
#> byvar2 = a   2 2.7683 [0.7819; 9.8007]     0   0
#> 
#> Test for subgroup differences (random effects model):
#>                   Q d.f. p-value
#> Between groups 0.00    0      --
#> 
#> Details on meta-analytical method:
#> - Mantel-Haenszel method (common effect model)
#> - Inverse variance method (random effects model)
#> - Restricted maximum-likelihood estimator for tau^2
#>