Define source order

Prioritize over multiple sources, according to some criteria

Usage

sort_sources(
  .data,
  method = c("source_list", "max_n"),
  source_name = NULL,
  source_list_order = NULL,
  study_id = NULL,
  study_n = NULL
)

Arguments

.data

A study- or event-level dataset @seealso e_data

method

Character string, one of "source_list", "max_n". See description.

source_name

Optional, character string, a column with sources

source_list_order

Optional, character vector, order of sources.

study_id

Optional, study identifier.

study_n

Optional, character string, a column with numbers (of patients).

Value

The original .data dataset, with an additional src numeric column.

Details

When multiples sources report safety data for the same event, they should be prioritized among each other, in order to retain the source with the highest priority for each event.

If source sorting has been decided use the source_list method. With this method, your dataset should have a column with the source_name, and you provide the order in the source_list_order argument. The first element of source_list_order is the highest priority source, the second element is the next highest priority, and so on.

If you want to sort sources according to highest number of patients in a study_arm (or any other number), use method max_n, and provide study_id and study_n arguments.

Examples

# Source order is already known, then use method "source_list"

 source_order <- c("src1", "src2")

 e_data |>
   sort_sources(
     method = c("source_list"),
     source_name = "source",
     source_list_order = source_order
   )
#>   source study_id              event_name n_event_arm1 n_arm1 src
#> 1   src1     NCT1   myocardial infarction            1     20   1
#> 2   src1     NCT1 acute coronary syndrome            2     20   1
#> 3   src1     NCT2   myocardial infarction            5     27   1
#> 4   src1     NCT2                  eczema           10     27   1
#> 5   src2     NCT2                  eczema            9     27   2
#> 6   src2     NCT2                pruritus            7     27   2
#> 7   src2     NCT3                pruritus           20     50   2

 # Sources with higher total number in arm1 have highest priority

   e_data_maxn <-
      e_data |>
        dplyr::select(study_id, n_arm1) |>
        dplyr::add_row(
          study_id = "NCT1",
          n_arm1   = 30
        )

  e_data_maxn |>
  sort_sources(
    method = c("max_n"),
    study_n = "n_arm1",
    study_id = "study_id"
    )
#>   study_id n_arm1 src
#> 1     NCT1     20   2
#> 2     NCT1     20   2
#> 3     NCT2     27   1
#> 4     NCT2     27   1
#> 5     NCT2     27   1
#> 6     NCT2     27   1
#> 7     NCT3     50   1
#> 8     NCT1     30   1