Table a model summary estimates

Shorthand to get results from one or more meta models as a data.frame

Usage

table_mod_summary(
  mod_list,
  main_or_sbgp = c("main", "sbgp"),
  common_or_random = c("common", "random"),
  prop_scaler = 1000
)

Arguments

mod_list

A named list of models, or a single meta model.

main_or_sbgp

Do you want to extract the main summary or subgroups?

common_or_random

Do you want common or random effects parameters?

prop_scaler

Only used when working with metaproportion or metarates, will express the result as x events per prop_scaler patients

Value

A data.frame with the following columns:

analysis The type of analysis (main outcome or subgroups)

mod_name The name of the model

event, event.e, event.c, n, n.e, n.c, k, I2, Q, pval.Q, Q.b, pval.Q.b: See ?meta::meta-object. For metaprop, Q, and pval.Q are from Wald test. For metarate models, Q and pval.Q are from Wald test.

n_stud_i2 Formating of I2 and k.

sm (95%CI) The summary effect and its 95% confidence interval ('sm' parameter)

method_sm_effects The method, summary measure and effect type

p The summary effect p-value

TE, lower, upper, either .common or .random as in ?meta::meta-object for main outcomes, and the same with .w for subgroups.

bylab The subgroup label

bylevs The subgroup levels

Details

The idea is to switch from a list of meta models, themselves lists, to an handy data.frame. At the moment, only works for metabin exports and metaprop "main" export. The function internally applies the core extraction to all items. A meta model does not incorporate much meta variables on what was summarized: the name of the list item is used and should be explicit enough to describe the whole setting (e.g. include adverse event name, type of data aggregation conservative/integrative, etc.)

Examples

library(meta)
#> Loading required package: metadat
#> Loading 'meta' package (version 8.0-1).
#> Type 'help(meta)' for a brief overview.

data <-
    data.frame(
      event = 4:1,
      n = c(5, 7, 8, 9),
      subgroup = c("a", "a", "b", "b")
    )

 m1 <- metaprop(
    event = event,
    n = n,
    subgroup = subgroup,
    data = data
  )

table_mod_summary(m1)
#>       analysis mod_name event  n n stud/I2     Proportion (95%CI)      per
#> 1 main outcome      mod    10 29     4/47% 344.83 (196.62-530.93) 1000 pts
#>    method_sm_effects        TE     lower     upper k        I2 Q..Wald..
#> 1 GLMM PLOGIT common 0.3448276 0.1966169 0.5309271 4 0.4681994  5.641212
#>   pval.Q.1.
#> 1 0.1304323

# If you have a list of models, named them
 res_mod <-
   table_mod_summary(
     list(mod1 = m1)
   )